+++ src/languages/textMate/index.ts
@@ -1,7 +1,7 @@
 import * as monaco from 'monaco-editor'
 import onigFile from 'vscode-oniguruma/release/onig.wasm'
-import { IGrammar, StackElement, IToken } from 'vscode-textmate'
 import CGTMGrammarFactory from './CGTMGrammarFactory'
+import CGTMTokenizationSupport from './CGTMTokenizationSupport'
 import rawGrammars from '../extensions/grammars.json'
 import grammarLoader from '../extensions/grammarLoader'
 
@@ -26,37 +26,6 @@ function getOrCreateGrammarFactory (): CGTMGrammarFactory {
     textMateGrammarFactory = createGrammarFactory()
   }
   return textMateGrammarFactory
-}
-
-function toMonacoToken (token: IToken) {
-  return {
-    ...token,
-    scopes: token.scopes.join(', ')
-  }
-}
-
-class CGTMTokenizationSupport extends monaco.extra.TMTokenizationSupport {
-  constructor (
-    private languageId: string,
-    encodedLanguageId: monaco.languages.LanguageId,
-    actual: monaco.extra.TMTokenization,
-    private grammar: IGrammar
-  ) {
-    super(languageId, encodedLanguageId, actual, monaco.editor.StaticServices.configurationService.get())
-  }
-
-  // To make "inspect tokens" work, default impl is `throw new Error('Not supported!');`
-  tokenize (line: string, hasEOL: boolean, state: monaco.languages.IState, offsetDelta: number): monaco.TokenizationResult {
-    return monaco.languages.adaptTokenize(this.languageId, {
-      tokenize: (line: string, state: monaco.languages.IState) => {
-        const actualResult = this.grammar.tokenizeLine(line, state as StackElement)
-        return {
-          tokens: actualResult.tokens.map(toMonacoToken),
-          endState: actualResult.ruleStack
-        }
-      }
-    }, line, state, offsetDelta)
-  }
 }
 
 const modeService = monaco.editor.StaticServices.modeService.get()
